Moving web services into a subdir.
