+++ IPython/frontend/html/notebook/services/clusters/handlers.py
@@ -20,7 +20,7 @@ from tornado import web
 
 from zmq.utils import jsonapi
 
-from ..base.handlers import IPythonHandler
+from ...base.handlers import IPythonHandler
 
 #-----------------------------------------------------------------------------
 # Cluster handlers
