+++ IPython/frontend/html/notebook/services/kernels/kernelmanager.py
