+++ IPython/frontend/html/notebook/services/notebooks/azurenbmanager.py
