+++ IPython/frontend/html/notebook/services/notebooks/nbmanager.py
